Fixed gettime syscall with sysenter wrapper issue

+++ SOA LAB - ZeOS/include/stats.h
@@ -1,6 +1,6 @@
 #ifndef STATS_H
 #define STATS_H
 
-extern unsigned long long zeos_ticks;
+extern unsigned int zeos_ticks;
 
 #endif /* !STATS_H */
